+++ kernel/multiboot.h
@@ -10,50 +10,52 @@
 #include <elf.h>
 #endif
 
+#include"types.h"
+
 /* See 'info multiboot': Specification->Boot info format */
 /* The memory map. Be careful that the offset 0 is `base_addr_low'
         instead of the `size' member. */
 typedef struct memory_map {
-    unsigned long size;
-    unsigned long base_addr_low;
-    unsigned long base_addr_high;
-    unsigned long length_low;
-    unsigned long length_high;
-    unsigned long type;
+    uint32 size;
+    uint32 base_addr_low;
+    uint32 base_addr_high;
+    uint32 length_low;
+    uint32 length_high;
+    uint32 type;
 } memory_map_t;
 
 typedef struct _multiboot_module {
     Elf32_Ehdr *pe;
     void *mod_end;
     char *string;
-    unsigned long reserved;
+    uint32 reserved;
 } multiboot_module;
 
 typedef struct _multiboot_drive {
-  unsigned long size;
+  uint32 size;
   unsigned char number;
   unsigned char mode;
-  unsigned short cylinders;
+  uint16 cylinders;
   unsigned char heads;
   unsigned char sectors;
-  unsigned short ports[];
+  uint16 ports[];
 } multiboot_drive;
 
 typedef struct _multiboot {
-    unsigned long flags;
-    unsigned long mem_lower;
-    unsigned long mem_upper;
-    unsigned long boot_device;
+    uint32 flags;
+    uint32 mem_lower;
+    uint32 mem_upper;
+    uint32 boot_device;
     char *cmdline;
-    unsigned long mods_count;
+    uint32 mods_count;
     multiboot_module *mods_addr;
-    unsigned long elf_num;
-    unsigned long elf_size;
+    uint32 elf_num;
+    uint32 elf_size;
     void *elf_addr;
     void *elf_shndx;
-    unsigned long mmap_length;
-    unsigned long mmap_addr;
-    unsigned long drives_length;
+    uint32 mmap_length;
+    uint32 mmap_addr;
+    uint32 drives_length;
     multiboot_drive *drives_addr;
     void *config_table;
     char *boot_loader_name;
